[qt] Fix MSVC C1202 by splitting ClientManager into focused TUs
MSVC fatal error C1202 ("recursive type or function dependency context
too complex") fires when rfl reflects too many types in a single
translation unit. ClientManager.cpp accumulated rfl::json::read
instantiations for IAM, session, HTTP, trading and eventing types in
one TU, hitting the constexpr depth limit in
rfl/internal/no_duplicate_field_names.hpp.

Split into four TUs, each carrying only the rfl instantiations relevant
to its domain:

- ClientManager.cpp         — connect/login/logout/session (IAM/HTTP types)
- ClientManagerSetup.cpp    — testConnection/signup (inline login/signup reads)
- ClientManagerTrades.cpp   — listTrades/getTradeDetail (trading types; 25-field
                              trade struct was the primary trigger)
- ClientManagerEvents.cpp   — subscribeToEvent/unsubscribeFromEvent
                              (entity_change_event)

The GLOB_RECURSE in src/CMakeLists.txt picks up the new files automatically.
No logic changes; all method signatures and behaviour are identical.
